pyproject.toml
[project]
name = "llm-parliament"
version = "0.1.0"
description = "An adversarial multi-agent debate platform."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "langgraph>=0.0.10",
    "langchain-openai>=0.0.5",
    "langchain-core>=0.1.10",
    "fastmcp>=0.1.0",
    "grandalf>=0.8"
]